requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
langchain-text-splitters
langchain-classic
pinecone